[RenderSystem] Implemented GLModel/GLMesh creation; Implemented Camera component; Refactor; Created view volume stubs

+++ include/render/Mesh.hpp
@@ -27,7 +27,7 @@ namespace zero::render {
          * @param vertices Interleaved vertices
          * @param indices Vertex indices
          */
-        Mesh(std::vector<Vertex>&& vertices, std::vector<uint32>&& indices);
+        Mesh(std::vector<Vertex>&& vertices, std::vector<uint16>&& indices);
 
         virtual ~Mesh() = default;
 
@@ -41,11 +41,11 @@ namespace zero::render {
          * @brief Get the vertex indices
          * @return the indices
          */
-        [[nodiscard]] const std::vector<uint32>& GetIndices() const;
+        [[nodiscard]] const std::vector<uint16>& GetIndices() const;
 
     protected:
         std::vector<Vertex> vertices_;
-        std::vector<uint32> indices_;
+        std::vector<uint16> indices_;
 
     }; // class Mesh
