+++ src/render/Mesh.cpp
@@ -2,7 +2,7 @@
 
 using namespace zero::render;
 
-Mesh::Mesh(std::vector<Vertex>&& vertices, std::vector<uint32>&& indices)
+Mesh::Mesh(std::vector<Vertex>&& vertices, std::vector<uint16>&& indices)
 : vertices_(std::move(vertices))
 , indices_(std::move(indices))
 {}
@@ -11,7 +11,6 @@ const std::vector<Vertex>& Mesh::GetVertices() const {
     return vertices_;
 }
 
-
-const std::vector<zero::uint32>& Mesh::GetIndices() const {
+const std::vector<zero::uint16>& Mesh::GetIndices() const {
     return indices_;
 }
